requirements.txt
# Dependencies

fastapi
uvicorn
groq
python-dotenv
pydantic